+++ packages/react-components/react-tags-preview/stories/InteractionTag/InteractionTagDismiss.stories.tsx
@@ -1,42 +1,34 @@
 import * as React from 'react';
-import { Avatar, makeStyles } from '@fluentui/react-components';
-import { CalendarMonthRegular } from '@fluentui/react-icons';
 
-import { InteractionTag, InteractionTagPrimary, InteractionTagSecondary } from '@fluentui/react-tags-preview';
+import {
+  InteractionTag,
+  InteractionTagPrimary,
+  InteractionTagSecondary,
+  TagGroup,
+  TagGroupProps,
+} from '@fluentui/react-tags-preview';
 
-const useContainerStyles = makeStyles({
-  root: {
-    display: 'flex',
-    flexDirection: 'column',
-    rowGap: '10px',
-  },
-});
+const initialTags = [
+  { value: '1', children: 'Tag 1' },
+  { value: '2', children: 'Tag 2' },
+  { value: '3', children: 'Tag 3' },
+];
 
 export const Dismiss = () => {
-  const containerStyles = useContainerStyles();
+  const [visibleTags, setVisibleTags] = React.useState(initialTags);
+  const removeItem: TagGroupProps['onDismiss'] = (_e, { dismissedTagValue }) => {
+    setVisibleTags([...visibleTags].filter(tag => tag.value !== dismissedTagValue));
+  };
+
   return (
-    <div className={containerStyles.root}>
-      <InteractionTag>
-        <InteractionTagPrimary hasSecondaryAction>Primary text</InteractionTagPrimary>
-        <InteractionTagSecondary />
-      </InteractionTag>
-      <InteractionTag>
-        <InteractionTagPrimary icon={<CalendarMonthRegular />} hasSecondaryAction>
-          Primary text
-        </InteractionTagPrimary>
-        <InteractionTagSecondary />
-      </InteractionTag>
-      <InteractionTag>
-        <InteractionTagPrimary
-          media={<Avatar name="Katri Athokas" badge={{ status: 'busy' }} />}
-          secondaryText="Secondary text"
-          hasSecondaryAction
-        >
-          Primary text
-        </InteractionTagPrimary>
-        <InteractionTagSecondary />
-      </InteractionTag>
-    </div>
+    <TagGroup onDismiss={removeItem}>
+      {visibleTags.map(tag => (
+        <InteractionTag value={tag.value} key={tag.value}>
+          <InteractionTagPrimary hasSecondaryAction>{tag.children}</InteractionTagPrimary>
+          <InteractionTagSecondary aria-label={`${tag.children}, remove`} />
+        </InteractionTag>
+      ))}
+    </TagGroup>
   );
 };
 
@@ -44,7 +36,8 @@ Dismiss.storyName = 'Dismiss';
 Dismiss.parameters = {
   docs: {
     description: {
-      story: 'An InteractionTag can have a button that dismisses it',
+      story:
+        'An InteractionTag can have a secondary action that is usually dismiss. TagGroup can handle dismiss for a collection of tags.',
     },
   },
 };